+++ db-queue-scheduler-core/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/DefaultScheduler.java
@@ -8,6 +8,7 @@ import ru.yoomoney.tech.dbqueue.scheduler.internal.schedule.NextExecutionTimePro
 import ru.yoomoney.tech.dbqueue.scheduler.internal.schedule.NextExecutionTimeProviderFactory;
 import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTask;
 import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskIdentity;
+import ru.yoomoney.tech.dbqueue.scheduler.models.info.ScheduledTaskInfo;
 import ru.yoomoney.tech.dbqueue.scheduler.settings.ScheduledTaskSettings;
 
 import javax.annotation.Nonnull;
@@ -30,7 +31,7 @@ class DefaultScheduler implements Scheduler {
 
     DefaultScheduler(@Nonnull ScheduledTaskManager scheduledTaskManager,
                      @Nonnull NextExecutionTimeProviderFactory nextExecutionTimeProviderFactory) {
-        this.scheduledTaskManager = requireNonNull(scheduledTaskManager, "scheduledTaskRegistry");
+        this.scheduledTaskManager = requireNonNull(scheduledTaskManager, "scheduledTaskManager");
         this.nextExecutionTimeProviderFactory = requireNonNull(nextExecutionTimeProviderFactory, "nextExecutionTimeProviderFactory");
     }
 
@@ -82,5 +83,10 @@ class DefaultScheduler implements Scheduler {
     @Override
     public List<ScheduledTaskIdentity> awaitTermination(@Nonnull Duration timeout) {
         return scheduledTaskManager.awaitTermination(timeout);
+    }
+
+    @Override
+    public List<ScheduledTaskInfo> getScheduledTaskInfo() {
+        return scheduledTaskManager.getScheduledTaskInfo();
     }
 }