+++ db-queue-scheduler-spring/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/SpringSchedulerConfigurator.java
@@ -143,6 +143,7 @@ public class SpringSchedulerConfigurator implements SchedulerConfigurator {
                 transactionOperations
         );
         ScheduledTaskQueueDao scheduledTaskQueueDao = new SpringScheduledTaskQueuePostgresDao(
+                tableName,
                 jdbcOperations,
                 transactionOperations,
                 QueueTableSchema.builder().build()