+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MINOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Scheduled tasks info feature implemented that allows listing all scheduled tasks info; 
 ### NEXT_VERSION_DESCRIPTION_END
 ## [1.0.0](https://github.com/yoomoney/db-queue-scheduler/pull/5) (29-10-2021)